refactor: move most shared logic to executor.rs
